suggestions

+++ src/app/content-holder/meal-plan/single-meal/single-meal.component.ts
@@ -4,7 +4,7 @@ import { MealPlanModalService } from './../meal-plan-modal.service';
 import { Router } from '@angular/router';
 import { ModalSingleMealComponent } from './../modal-single-meal/modal-single-meal.component';
 import { FoodRow } from './../../../shared/models/food-row.model';
-import { Component, OnInit, Input, Output, EventEmitter} from '@angular/core';
+import { Component, OnInit, Input, Output, EventEmitter } from '@angular/core';
 import { SingleMeal } from '../../../shared/models/single-meal.model';
 
 
@@ -18,7 +18,7 @@ import { BsModalService } from 'ngx-bootstrap/modal';
 })
 export class SingleMealComponent implements OnInit {
 
-  @Input() singleMeal: SingleMeal;  
+  @Input() singleMeal: SingleMeal;
   @Input() clickedUsername: string;
   @Input() dateFromDatepicker: string;
   @Input() trainerClientStatus: string;
@@ -28,12 +28,12 @@ export class SingleMealComponent implements OnInit {
   public isMyClientsClicked: boolean;
 
   constructor(
-    private router: Router ,
+    private router: Router,
     private modalService: MealPlanModalService,
     private mealHttpService: MealHttpService,
     private utilsService: UtilsService
-  ) { 
-    this.isMyClientsClicked=this.utilsService.isMyClientsClicked(this.router.url);
+  ) {
+    this.isMyClientsClicked = this.utilsService.isMyClientsClicked(this.router.url);
     this.onDelete = new EventEmitter();
   }
 
@@ -41,21 +41,23 @@ export class SingleMealComponent implements OnInit {
     let url = this.router.url;
   }
 
-  openModal(){
+  openModal() {
     this.modalService.openModal(this.singleMeal, this.clickedUsername, this.dateFromDatepicker);
   }
 
-  deleteSingleMeal(){
+  deleteSingleMeal() {
 
     let clientUsername = this.clickedUsername;
     let trainerUsername = localStorage.getItem("username");
 
-    this.mealHttpService.deleteSingleMeal(this.singleMeal.id, clientUsername,  trainerUsername).subscribe(
-      data => {
-        console.log("deleted");
-        this.onDelete.emit(this.singleMeal.id);
-      }
-    );
+    if (window.confirm("Are you sure you want to delete this Meal?")) {
+      this.mealHttpService.deleteSingleMeal(this.singleMeal.id, clientUsername, trainerUsername).subscribe(
+        data => {
+          console.log("deleted");
+          this.onDelete.emit(this.singleMeal.id);
+        }
+      );
+    }
   }
 
 }
